+++ autometrics/src/tracker/opentelemetry.rs
@@ -1,13 +1,33 @@
 use crate::labels::{BuildInfoLabels, CounterLabels, GaugeLabels, HistogramLabels, Label};
 use crate::{constants::*, tracker::TrackMetrics};
-use opentelemetry_api::{global, metrics::UpDownCounter, Context, KeyValue};
+use once_cell::sync::Lazy;
+use opentelemetry_api::metrics::{Counter, Histogram, UpDownCounter};
+use opentelemetry_api::{global, Context, KeyValue};
 use std::{sync::Once, time::Instant};
 
 static SET_BUILD_INFO: Once = Once::new();
+static COUNTER: Lazy<Counter<u64>> = Lazy::new(|| {
+    global::meter("")
+        .u64_counter(COUNTER_NAME)
+        .with_description(COUNTER_DESCRIPTION)
+        .init()
+});
+static HISTOGRAM: Lazy<Histogram<f64>> = Lazy::new(|| {
+    global::meter("")
+        .f64_histogram(HISTOGRAM_NAME)
+        .with_description(HISTOGRAM_DESCRIPTION)
+        .init()
+});
+static GAUGE: Lazy<UpDownCounter<i64>> = Lazy::new(|| {
+    global::meter("")
+        .i64_up_down_counter(GAUGE_NAME)
+        .with_description(GAUGE_DESCRIPTION)
+        .init()
+});
 
 /// Tracks the number of function calls, concurrent calls, and latency
 pub struct OpenTelemetryTracker {
-    concurrency_tracker: Option<(UpDownCounter<i64>, Vec<KeyValue>)>,
+    gauge_labels: Option<Vec<KeyValue>>,
     start: Instant,
     context: Context,
 }
@@ -16,21 +36,17 @@ impl TrackMetrics for OpenTelemetryTracker {
     fn start(gauge_labels: Option<&GaugeLabels>) -> Self {
         let context = Context::current();
 
-        let concurrency_tracker = if let Some(gauge_labels) = gauge_labels {
+        let gauge_labels = if let Some(gauge_labels) = gauge_labels {
             let gauge_labels = to_key_values(gauge_labels.to_array());
             // Increase the number of concurrent requests
-            let concurrency_tracker = global::meter("")
-                .i64_up_down_counter(GAUGE_NAME)
-                .with_description(GAUGE_DESCRIPTION)
-                .init();
-            concurrency_tracker.add(&context, 1, &gauge_labels);
-            Some((concurrency_tracker, gauge_labels))
+            GAUGE.add(&context, 1, &gauge_labels);
+            Some(gauge_labels)
         } else {
             None
         };
 
         Self {
-            concurrency_tracker,
+            gauge_labels,
             start: Instant::now(),
             context,
         }
@@ -40,24 +56,16 @@ impl TrackMetrics for OpenTelemetryTracker {
         let duration = self.start.elapsed().as_secs_f64();
 
         // Track the function calls
-        let counter = global::meter("")
-            .f64_counter(COUNTER_NAME)
-            .with_description(COUNTER_DESCRIPTION)
-            .init();
         let counter_labels = to_key_values(counter_labels.to_vec());
-        counter.add(&self.context, 1.0, &counter_labels);
+        COUNTER.add(&self.context, 1, &counter_labels);
 
         // Track the latency
-        let histogram = global::meter("")
-            .f64_histogram(HISTOGRAM_NAME)
-            .with_description(HISTOGRAM_DESCRIPTION)
-            .init();
         let histogram_labels = to_key_values(histogram_labels.to_vec());
-        histogram.record(&self.context, duration, &histogram_labels);
+        HISTOGRAM.record(&self.context, duration, &histogram_labels);
 
         // Decrease the number of concurrent requests
-        if let Some((concurrency_tracker, gauge_labels)) = self.concurrency_tracker {
-            concurrency_tracker.add(&self.context, -1, &gauge_labels);
+        if let Some(gauge_labels) = self.gauge_labels {
+            GAUGE.add(&self.context, -1, &gauge_labels);
         }
     }