+++ autometrics/src/tracker/prometheus.rs
@@ -1,6 +1,5 @@
-use crate::labels::{BuildInfoLabels, CounterLabels, GaugeLabels, HistogramLabels};
+use crate::labels::{BuildInfoLabels, CounterLabels, GaugeLabels, HistogramLabels, ResultLabel};
 use crate::{constants::*, tracker::TrackMetrics, HISTOGRAM_BUCKETS};
-use const_format::{formatcp, str_replace};
 use once_cell::sync::Lazy;
 use prometheus::core::{AtomicI64, GenericGauge};
 use prometheus::{
@@ -8,13 +7,6 @@ use prometheus::{
     HistogramVec, IntCounterVec, IntGaugeVec,
 };
 use std::{sync::Once, time::Instant};
-
-const COUNTER_NAME_PROMETHEUS: &str = str_replace!(COUNTER_NAME, ".", "_");
-const HISTOGRAM_NAME_PROMETHEUS: &str = str_replace!(HISTOGRAM_NAME, ".", "_");
-const GAUGE_NAME_PROMETHEUS: &str = str_replace!(GAUGE_NAME, ".", "_");
-const OBJECTIVE_NAME_PROMETHEUS: &str = str_replace!(OBJECTIVE_NAME, ".", "_");
-const OBJECTIVE_PERCENTILE_PROMETHEUS: &str = str_replace!(OBJECTIVE_PERCENTILE, ".", "_");
-const OBJECTIVE_LATENCY_PROMETHEUS: &str = str_replace!(OBJECTIVE_LATENCY_THRESHOLD, ".", "_");
 
 static SET_BUILD_INFO: Once = Once::new();
 
@@ -33,7 +25,7 @@ static COUNTER: Lazy<IntCounterVec> = Lazy::new(|| {
             OBJECTIVE_PERCENTILE_PROMETHEUS,
         ]
     )
-    .expect(formatcp!(
+    .expect(&format!(
         "Failed to register {COUNTER_NAME_PROMETHEUS} counter"
     ))
 });
@@ -54,7 +46,7 @@ static HISTOGRAM: Lazy<HistogramVec> = Lazy::new(|| {
             MODULE_KEY,
             OBJECTIVE_NAME_PROMETHEUS,
             OBJECTIVE_PERCENTILE_PROMETHEUS,
-            OBJECTIVE_LATENCY_PROMETHEUS
+            OBJECTIVE_LATENCY_THRESHOLD_PROMETHEUS
         ]
     )
     .expect("Failed to register function_calls_duration histogram")
@@ -107,27 +99,19 @@ impl TrackMetrics for PrometheusTracker {
                     counter_labels.function,
                     counter_labels.module,
                     counter_labels.caller,
-                    counter_labels.result.unwrap_or_default().0,
-                    if let Some((OK_KEY, Some(return_value_type))) = counter_labels.result {
-                        return_value_type
-                    } else {
-                        ""
+                    match counter_labels.result {
+                        Some(ResultLabel::Ok) => OK_KEY,
+                        Some(ResultLabel::Error) => ERROR_KEY,
+                        None => "",
                     },
-                    if let Some((ERROR_KEY, Some(return_value_type))) = counter_labels.result {
-                        return_value_type
-                    } else {
-                        ""
-                    },
+                    counter_labels.ok.unwrap_or_default(),
+                    counter_labels.error.unwrap_or_default(),
+                    counter_labels.objective_name.unwrap_or_default(),
                     counter_labels
-                        .objective
+                        .objective_percentile
                         .as_ref()
-                        .map(|obj| obj.0)
-                        .unwrap_or(""),
-                    counter_labels
-                        .objective
-                        .as_ref()
-                        .map(|obj| obj.1.as_str())
-                        .unwrap_or(""),
+                        .map(|p| p.as_str())
+                        .unwrap_or_default(),
                 ],
             )
             .inc();
@@ -136,21 +120,17 @@ impl TrackMetrics for PrometheusTracker {
             .with_label_values(&[
                 histogram_labels.function,
                 histogram_labels.module,
+                histogram_labels.objective_name.unwrap_or_default(),
                 histogram_labels
-                    .objective
+                    .objective_percentile
                     .as_ref()
-                    .map(|obj| obj.0)
-                    .unwrap_or(""),
+                    .map(|p| p.as_str())
+                    .unwrap_or_default(),
                 histogram_labels
-                    .objective
+                    .objective_latency_threshold
                     .as_ref()
-                    .map(|obj| obj.1.as_str())
-                    .unwrap_or(""),
-                histogram_labels
-                    .objective
-                    .as_ref()
-                    .map(|obj| obj.2.as_str())
-                    .unwrap_or(""),
+                    .map(|p| p.as_str())
+                    .unwrap_or_default(),
             ])
             .observe(duration);